requirements.txt
groq>=1.0.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
langchain==0.1.9
langchain-community==0.0.24
